6 User Create,Show Questions-Answers

+++ app/controllers/questions_controller.rb
@@ -1,33 +1,20 @@
 class QuestionsController < ApplicationController
 
   before_action :authenticate_user!, except: %i[index show]  
-  # before_action :load_question, only: [:show, :edit, :update, :destroy]
   
   expose :questions, ->{ Question.all }
   expose :question
-
-  # def index
-  #   @questions = Question.all
-  # end
-
-  # def show
-  # end
-
-  # def new
-  #   @question = Question.new
-  # end
-
-  # def edit
-  # end
+  # expose :answer, -> { Answer.new }
 
   def create
     # create - сохранение сразу в базу а нужно условное сохранение
-		# @question = Question.create(question_params)
-    
-    @question = Question.new(question_params)
+		# @question = Question.create(question_params)    
+    # @question = Question.new(question_params)
+
+    @exposed_question = current_user.author_questions.new(question_params)
 
     if question.save
-      redirect_to @question, notice: 'Your question successfully created.'
+      redirect_to question_path(question), notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -42,18 +29,17 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question was successfully deleted.'
+    else
+      redirect_to question, notice: 'You have no rights to do this.'
+    end
   end
 
   private
 
-  # def load_question
-  #   @question = Question.find(params[:id])
-  # end
-
   def question_params
     params.require(:question).permit(:title, :body)
   end
-
 end
